EWPP-1468: Consolidate implementation of event subscriber regarding extractling links for each author.
