--- modules/oe_content_sub_entity/modules/oe_content_sub_entity_extra_authors/oe_content_sub_entity_extra_authors.services.yml
@@ -1,6 +1,0 @@
-services:
-  oe_content_sub_entity_extra_authors.author_event_subscriber:
-    class: Drupal\oe_content_sub_entity_extra_authors\EventSubscriber\AuthorExtraBundlesSubscriber
-    parent: oe_content.author_event_subscriber.base
-    tags:
-      - { name: event_subscriber }
